% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_IUPred2A_results_csv.R
\name{write_IUPred2A_results_csv}
\alias{write_IUPred2A_results_csv}
\title{Write a .results File from the IUPred2A Webservice as a Data Frame}
\usage{
write_IUPred2A_results_csv(inFile, outFile, colNameRow = 5L, delim = "\\t")
}
\arguments{
\item{inFile}{File path to a \code{.results} file saved on your local machine}

\item{outFile}{File path to write the \code{.csv} file}

\item{colNameRow}{Which row has the column names? Defaults to 5, but this
may need to be adjusted if the \code{.results} file format changes.}

\item{delim}{Within each protein, what is the results delimiter? Defaults to
\code{"\\t"}, which indicates the tablular data are tab delimited.}
}
\value{
Nothing. Writes a \code{.csv} file with columns for the FASTA header
and IUPred2A output (such as the position, amino acid, and discordance
scores)
}
\description{
Given a file of disordered protein regions results from the
IUPred2A online calculator, write these results to a \code{.csv} file
}
\details{
This function works in 4 main chunks: 1) import the data and extract
the (future) column names; 2) split the character strings (one per row)
into groups based on the protein (FASTA name); 3) transform these groups of
character strings into a character matrix, one per protein; 4) add back
the protein names and column names and "stack" all matrices into one
}
\examples{
\dontrun{
  write_IUPred2A_results_csv(
    inFile  = "inst/extdata/data_raw/COMBINED_BLAST2_MUSCLE.result",
    outFile = "inst/extdata/data_clean/muscle_BLAST2_results_20220422.csv"
  )
}
  
}
